// ClientApp/src/Types/ICountries.ts
export default interface ICountries {
  countryID: number;
  languageID: number;
  name: string;
  countryOfficialLanguageID: number;
  alpha3Code: string;
  currencyCodes: string;
  amountToPay: number;
  equivalenceIn$: number;
  phoneNumberCode: number;
  currencyName: string;
  currency_iso: string;
  imageLink: string;
  name_iso2: string;
  name_iso3: string;
  phonenumberCode: string;
}